+++ lib/src/video_player_initializer/network_initializer.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/cupertino.dart';
+import 'package:omni_video_player/src/api/hls_video_api.dart';
 import 'package:omni_video_player/src/controllers/default_playback_controller.dart';
 import 'package:omni_video_player/src/video_player_initializer/video_player_initializer_factory.dart';
 import 'package:omni_video_player/omni_video_player.dart';
@@ -20,8 +21,25 @@ class NetworkInitializer implements IVideoPlayerInitializerStrategy {
   @override
   Future<OmniPlaybackController?> initialize() async {
     try {
+      final isHlsVideo = await HlsVideoApi.isHlsUri(
+          options.videoSourceConfiguration.videoUrl!);
+
+      Map<OmniVideoQuality, Uri>? qualitiesMap;
+      MapEntry<OmniVideoQuality, Uri>? currentQualityEntry;
+
+      if (isHlsVideo) {
+        qualitiesMap = await HlsVideoApi.extractHlsVariantsByQuality(
+            options.videoSourceConfiguration.videoUrl!,
+            options.videoSourceConfiguration.availableQualities);
+        currentQualityEntry = HlsVideoApi.selectBestQualityVariant(qualitiesMap,
+            preferredQualities:
+                options.videoSourceConfiguration.preferredQualities);
+      }
+
       final controller = await DefaultPlaybackController.create(
-        videoUrl: options.videoSourceConfiguration.videoUrl!,
+        videoUrl: (currentQualityEntry != null)
+            ? currentQualityEntry.value
+            : options.videoSourceConfiguration.videoUrl!,
         dataSource: null,
         audioUrl: null,
         isLive: false,
@@ -31,6 +49,9 @@ class NetworkInitializer implements IVideoPlayerInitializerStrategy {
         callbacks: callbacks,
         type: options.videoSourceConfiguration.videoSourceType,
         globalKeyPlayer: options.globalKeyPlayer,
+        qualityUrls: qualitiesMap,
+        currentVideoQuality:
+            (currentQualityEntry != null) ? currentQualityEntry.key : null,
       );
 
       controller.sharedPlayerNotifier.value = Hero(